The fruit of today's labor, getting "fb:" salt in place for newly created users
This makes sure that we don't have other systems stomping on each other since
it's entirely conceivable to have multiple systems with a user 123321.  If we
store just the Facebook uid field, we have troubles when we try to store a uid
from Twitter, or Google, or Ning, etc., etc.  This starts down the correct path
to fix that problem in the future.