Refactor the entire structure of this. (major BC break)
With two pieces in here, it makes sense to treat them like two separate apps
and pull them in as necessary.  This moves everything around to do that.  Still
a bit in flux, however, as I may simplify some of the class names now that they
are properly namespaced.

This does break BC for anyone who's depending on this code.  Since that should
still be just me, not a big issue, but you've been warned. :-)